// ntis-web/projects/itree-web/src/app/wte-maintenance-and-operation/contracts/enums/contract-state.enums.ts
export enum ContractState {
  SUBMITTED = 'SUBMITTED',
  CANCELLED = 'CANCELLED',
  REJECTED = 'REJECTED',
  CONFIRMED = 'CONFIRMED',
  TERMINATED = 'TERMINATED',
  SIGNED_BY_BOTH = 'SIGNED_BY_BOTH',
  SIGNED_BY_SRV_PROVIDER = 'SIGNED_BY_SRV_PROV',
  SIGNED_BY_CUSTOMER = 'SIGNED_BY_CUSTOMER',
  VALID = 'VALID',
  EXPIRED = 'EXPIRED',
}